# [statedb] clear snapshots after running each transaction