when workingset is running all tx in the block, all snapshots made during one tx are meaningless for the next tx, so after running one tx, clear all snapshots to reduce memory consumption/growth of underlying `CachedBatch`