@dustinxie 2022-01-27
> why we should make this change if the current logic works fine and makes sense?

1. this logic is more correct (snapshots from previous tx is useless for next tx) 
2. reduce memory usage, increase perf