@dustinxie 2022-03-03
this change reduced memory usage while running actions, also does not hurt running time
before
```
BenchmarkMintAndCommitBlock-8   	       1	1050992664 ns/op
BenchmarkMintBlock-8            	                      2	 509487252 ns/op
BenchmarkValidateBlock-8        	                      1	1166334819 ns/op
```
after
```
BenchmarkMintAndCommitBlock-8   	       1	1031465595 ns/op
BenchmarkMintBlock-8            	                      2	 518841254 ns/op
BenchmarkValidateBlock-8        	                      1	1154082540 ns/op
```