@dustinxie 2022-03-03
> why we should make this change if the current logic works fine and makes sense?

IMO this change removes unnecessary snapshots and reduces run-time memory, the logic still makes sense with this change